move binding generation out of aws-lc-sys
