+++ bindings/rust/aws-lc-sys-template/tests/sanity-tests.rs
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0 OR ISC
-use std::mem::MaybeUninit;
 use aws_lc_sys;
+use std::mem::MaybeUninit;
 
 use openssl;
