+++ bindings/rust/generate/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "generate"
+version = "__AWS_LC_SYS_VERSION__"
+authors = ["AWS-LC"]
+edition = "2021"
+license = "Apache-2.0 OR ISC"
+# this is an internal tool for generating bindings
+publish = false
+
+[dependencies]
+bindgen = "0.61"
+dunce = "1.0"
